Remove custom curried impls
