--- clvm-macros/src/crate_ident.rs
@@ -1,15 +1,0 @@
-use proc_macro2::{Ident, Span, TokenStream};
-use proc_macro_crate::{crate_name, FoundCrate};
-use quote::quote;
-
-pub fn crate_ident() -> TokenStream {
-    let found_crate = crate_name("clvm-utils").expect("`clvm-utils` not found in `Cargo.toml`");
-
-    match found_crate {
-        FoundCrate::Itself => quote!(crate),
-        FoundCrate::Name(name) => {
-            let ident = Ident::new(&name, Span::call_site());
-            quote!(#ident)
-        }
-    }
-}
